exported stats table to external component
